requirements.txt
psycopg2-binary
pandas
pandas-datareader
Pyarrow
yfinance
streamlit